refactor(anthropic): split client tests for GH727
